Refactor; if 1 open claim exists, instantly navigate to it
